Refactor: 인터셉터, 리졸버 위치 변경
- AsIs: global.interceptor, global.resolver 아래
- ToBe: 각 authentication, authorization 패키지 아래
