+++ src/main/java/com/uranus/taskmanager/api/security/authentication/interceptor/LoginRequired.java
@@ -1,14 +1,11 @@
-package com.uranus.taskmanager.api.global.interceptor;
+package com.uranus.taskmanager.api.security.authentication.interceptor;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import com.uranus.taskmanager.api.workspacemember.WorkspaceRole;
-
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface RoleRequired {
-	WorkspaceRole[] roles();
+public @interface LoginRequired {
 }
