+++ src/main/java/com/uranus/taskmanager/api/security/authentication/presentation/controller/AuthenticationController.java
@@ -9,12 +9,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.uranus.taskmanager.api.global.interceptor.LoginRequired;
+import com.uranus.taskmanager.api.common.ApiResponse;
 import com.uranus.taskmanager.api.security.authentication.constant.SessionKey;
+import com.uranus.taskmanager.api.security.authentication.interceptor.LoginRequired;
 import com.uranus.taskmanager.api.security.authentication.presentation.dto.request.LoginRequest;
 import com.uranus.taskmanager.api.security.authentication.presentation.dto.response.LoginResponse;
 import com.uranus.taskmanager.api.security.authentication.service.AuthenticationService;
-import com.uranus.taskmanager.api.common.ApiResponse;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
@@ -35,13 +35,7 @@ public class AuthenticationController {
 		LoginResponse loginResponse = authenticationService.login(loginRequest);
 
 		HttpSession session = request.getSession();
-		/*
-		 * Todo
-		 *  - 다음을 고려 중
-		 *  - MemberSession을 만들어서 객체를 세션에 저장할까? -> 직렬화 필요!
-		 *  - 지금 처럼 사용할 가능성이 높은 정보를 모두 저장할까? -> 특정 경우에 id를 통해 멤버를 찾는 과정을 생략할 수 있음
-		 *  - 그냥 id만 저장할까?
-		 */
+
 		session.setAttribute(SessionKey.LOGIN_MEMBER_ID, loginResponse.getId());
 		session.setAttribute(SessionKey.LOGIN_MEMBER_LOGIN_ID, loginResponse.getLoginId());
 		session.setAttribute(SessionKey.LOGIN_MEMBER_EMAIL, loginResponse.getEmail());
