+++ src/main/java/com/uranus/taskmanager/api/security/authorization/interceptor/RoleRequired.java
@@ -1,11 +1,14 @@
-package com.uranus.taskmanager.api.global.interceptor;
+package com.uranus.taskmanager.api.security.authorization.interceptor;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import com.uranus.taskmanager.api.workspacemember.WorkspaceRole;
+
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LoginRequired {
+public @interface RoleRequired {
+	WorkspaceRole[] roles();
 }
