+++ src/test/java/com/uranus/taskmanager/api/authentication/ResolveLoginMemberArgumentResolverTest.java
@@ -15,14 +15,14 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.context.request.NativeWebRequest;
 
-import com.uranus.taskmanager.api.global.resolver.LoginMemberArgumentResolver;
-import com.uranus.taskmanager.api.global.resolver.ResolveLoginMember;
-import com.uranus.taskmanager.api.security.authentication.constant.SessionKey;
-import com.uranus.taskmanager.api.security.authentication.presentation.dto.LoginMember;
-import com.uranus.taskmanager.api.security.authentication.exception.UserNotLoggedInException;
 import com.uranus.taskmanager.api.member.domain.Member;
-import com.uranus.taskmanager.api.member.exception.MemberNotFoundException;
 import com.uranus.taskmanager.api.member.domain.repository.MemberRepository;
+import com.uranus.taskmanager.api.member.exception.MemberNotFoundException;
+import com.uranus.taskmanager.api.security.authentication.constant.SessionKey;
+import com.uranus.taskmanager.api.security.authentication.exception.UserNotLoggedInException;
+import com.uranus.taskmanager.api.security.authentication.presentation.dto.LoginMember;
+import com.uranus.taskmanager.api.security.authentication.resolver.LoginMemberArgumentResolver;
+import com.uranus.taskmanager.api.security.authentication.resolver.ResolveLoginMember;
 import com.uranus.taskmanager.fixture.entity.MemberEntityFixture;
 
 import jakarta.servlet.http.HttpServletRequest;
