+++ src/test/java/com/uranus/taskmanager/api/workspacemember/authorization/AuthorizationInterceptorTest.java
@@ -16,13 +16,13 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.web.method.HandlerMethod;
 
-import com.uranus.taskmanager.api.global.interceptor.AuthorizationInterceptor;
-import com.uranus.taskmanager.api.global.interceptor.RoleRequired;
 import com.uranus.taskmanager.api.member.domain.Member;
 import com.uranus.taskmanager.api.security.authentication.constant.SessionKey;
 import com.uranus.taskmanager.api.security.authentication.exception.UserNotLoggedInException;
 import com.uranus.taskmanager.api.security.authorization.exception.InsufficientWorkspaceRoleException;
 import com.uranus.taskmanager.api.security.authorization.exception.InvalidWorkspaceCodeInUriException;
+import com.uranus.taskmanager.api.security.authorization.interceptor.AuthorizationInterceptor;
+import com.uranus.taskmanager.api.security.authorization.interceptor.RoleRequired;
 import com.uranus.taskmanager.api.workspace.domain.Workspace;
 import com.uranus.taskmanager.api.workspace.domain.repository.WorkspaceRepository;
 import com.uranus.taskmanager.api.workspace.exception.WorkspaceNotFoundException;
